Moved CodeWithPayloadError to agent package since not required in api anymore
