+++ api/api.go
@@ -398,18 +398,6 @@ func defaultConfig(transportFn func() *http.Transport) *Config {
 	return config
 }
 
-// CodeWithPayloadError allow returning non HTTP 200
-// Error codes while not returning PlainText payload
-type CodeWithPayloadError struct {
-	Reason      string
-	StatusCode  int
-	ContentType string
-}
-
-func (e CodeWithPayloadError) Error() string {
-	return e.Reason
-}
-
 // TLSConfig is used to generate a TLSClientConfig that's useful for talking to
 // Consul using TLS.
 func SetupTLSConfig(tlsConfig *TLSConfig) (*tls.Config, error) {
@@ -785,11 +773,9 @@ func (c *Client) doRequest(r *request) (time.Duration, *http.Response, error) {
 func (c *Client) query(endpoint string, out interface{}, q *QueryOptions) (*QueryMeta, error) {
 	r := c.newRequest("GET", endpoint)
 	r.setQueryOptions(q)
-	rtt, resp, err := requireOK(c.doRequest(r))
+	rtt, resp, err := c.doRequest(r)
 	if err != nil {
-		if _, ok := err.(CodeWithPayloadError); !ok {
-			return nil, err
-		}
+		return nil, err
 	}
 	defer resp.Body.Close()
